requirements.txt
openai
langchain
python-dotenv
langchain-openai
zhipuai
langchain-community
chromadb
streamlit
pymupdf
watchdog